requirements.txt
streamlit
python-pptx
groq
openai
anthropic
google-generativeai